use additionalMenuItemOnTopRenderer
