+++ apps/app/src/components/PageList/BookmarkList.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useState } from 'react';
+import React, { useCallback, useState, useMemo } from 'react';
 
 import nodePath from 'path';
 
@@ -6,13 +6,15 @@ import {
   IPageInfoAll, IPageToDeleteWithMeta, pathUtils,
 } from '@growi/core';
 import { useTranslation } from 'next-i18next';
-import { DropdownToggle } from 'reactstrap';
+import { DropdownItem, DropdownToggle } from 'reactstrap';
 
 import { unbookmark } from '~/client/services/page-operation';
 import { apiv3Put } from '~/client/util/apiv3-client';
+import { addBookmarkToFolder } from '~/client/util/bookmark-utils';
 import { ValidationTarget } from '~/client/util/input-validator';
 import { toastError, toastSuccess } from '~/client/util/toastr';
 import { IPageHasId } from '~/interfaces/page';
+import { useSWRxCurrentUserBookmarks } from '~/stores/bookmark';
 import loggerFactory from '~/utils/logger';
 
 import ClosableTextInput from '../Common/ClosableTextInput';
@@ -34,6 +36,36 @@ export const BookmarkList = (props:Props): JSX.Element => {
   } = props;
   const { t } = useTranslation();
   const [isRenameInputShown, setIsRenameInputShown] = useState(false);
+
+  const pageId = page._id;
+
+  const { data: userBookmarks, mutate: mutateUserBookmarks } = useSWRxCurrentUserBookmarks();
+  const isMoveToRoot = useMemo(() => {
+    return !userBookmarks?.map(userBookmark => userBookmark._id).includes(pageId);
+  }, [pageId, userBookmarks]);
+
+  const moveToRootClickedHandler = useCallback(async() => {
+    try {
+      await addBookmarkToFolder(pageId, null);
+      await mutateUserBookmarks();
+    }
+    catch (err) {
+      toastError(err);
+    }
+  }, [mutateUserBookmarks, pageId]);
+
+  const additionalMenuItemOnTopRenderer = useMemo(() => {
+    return (
+      <DropdownItem
+        onClick={moveToRootClickedHandler}
+        className="grw-page-control-dropdown-item"
+        data-testid="add-remove-bookmark-btn"
+      >
+        <i className="fa fa-fw fa-bookmark-o grw-page-control-dropdown-icon"></i>
+        {t('bookmark_folder.move_to_root')}
+      </DropdownItem>
+    );
+  }, [moveToRootClickedHandler, t]);
 
   const bookmarkMenuItemClickHandler = useCallback(async() => {
     await unbookmark(page._id);
@@ -103,6 +135,7 @@ export const BookmarkList = (props:Props): JSX.Element => {
         onClickBookmarkMenuItem={bookmarkMenuItemClickHandler}
         onClickRenameMenuItem={() => setIsRenameInputShown(true)}
         onClickDeleteMenuItem={deleteMenuItemClickHandler}
+        additionalMenuItemOnTopRenderer={isMoveToRoot ? (() => additionalMenuItemOnTopRenderer) : undefined}
       >
         <DropdownToggle color="transparent" className="border-0 rounded btn-page-item-control p-0 grw-visible-on-hover mr-1">
           <i className="icon-options fa fa-rotate-90 p-1"></i>
